llm_council/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-council"
version = "1.0.0"
description = "Multi-model AI consensus system - Get validated answers from Claude, GPT-4, Gemini, and more"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Dima", email = "dima@example.com"}
]
maintainers = [
    {name = "Dima"}
]
keywords = [
    "llm",
    "ai",
    "consensus",
    "multi-model",
    "claude",
    "gpt-4",
    "gemini",
    "grok",
    "synthesis",
    "voting"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.34.0",
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
web = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]
all = [
    "llm-council[dev,web]",
]

[project.urls]
Homepage = "https://github.com/Bafff/claude-code-council"
Repository = "https://github.com/Bafff/claude-code-council"
Issues = "https://github.com/Bafff/claude-code-council/issues"
Documentation = "https://github.com/Bafff/claude-code-council/tree/main/llm-council"

[project.scripts]
llm-council = "llm_council.cli:main"
council = "llm_council.cli:main"

[tool.setuptools]
packages = ["adapters", "core", "interfaces"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.md"]

# Black code formatter
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

# Ruff linter
[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__ are ok

# Pytest
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"

# MyPy type checker
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

# Coverage
[tool.coverage.run]
source = ["adapters", "core", "interfaces"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
